--- lib/src/main/java/io/github/thoroldvix/internal/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package io.github.thoroldvix.internal;
-
-import org.jspecify.annotations.NullMarked;
